backend/requirements.txt
# File: backend/requirements.txt
#
# PURPOSE
# -------
# Python dependencies for the backend service.
#
# DESIGN RULES
# ------------
# - Pinned versions for reproducibility
# - Async-first stack
# - Production-grade observability & security
# - No dev/test-only deps here (tests are paused)

# -------------------------------
# Web framework & server
# -------------------------------
fastapi==0.110.0
uvicorn[standard]==0.29.0

# -------------------------------
# Async DB & ORM
# -------------------------------
SQLAlchemy==2.0.29
asyncpg==0.29.0

# -------------------------------
# Auth & security
# -------------------------------
python-jose==3.3.0
cryptography==42.0.5

# -------------------------------
# Messaging
# -------------------------------
kafka-python==2.0.2

# -------------------------------
# Data validation & utilities
# -------------------------------
pydantic==2.6.4
python-dotenv==1.0.1

# -------------------------------
# Observability
# -------------------------------
loguru==0.7.2

# -------------------------------
# HTTP clients
# -------------------------------
httpx==0.27.0

# -------------------------------
# AI / inference bridge (Python side)
# -------------------------------
numpy==1.26.4

# -------------------------------
# Optional (enable when needed)
# -------------------------------
# redis==5.0.3
# sentry-sdk==1.44.0
PyJWT==2.8.0
